Fix 1-90 call audio upload SQL error and Gameplay Announcements overflow
uploadNumberAudio()'s upsert query used $2 both as the integer `number`
column and inside 'Number ' || $2 (implicit text concatenation) in the
same statement. Postgres infers one type per placeholder across an
entire query, so this always threw "inconsistent types deduced for
parameter $2: text versus integer" (42P08) - a real 500 on every
upload, both languages. It looked ENG-specific only because NEP's
existing files came from migration 042's backfill UPDATE, not this
INSERT path, so NEP was never actually exercised here. The file write to
disk happens before this query runs, which is why playback still worked
despite the error - the DB row update (audio_url_en) silently never
happened; only the fallback /audio/calls/{n}_en.mp3 convention path in
useGameAudio.ts made it play at all. Fixed with an explicit $2::int cast
everywhere $2 appears. Verified via a live upload against the local
backend: 500 -> 200, audio_url_en correctly returned and persisted.

Also fixed the Gameplay Announcements (Intro/Outro) English/Nepali
sub-cards, which used a fixed `1fr 1fr` grid with no minimum-width
floor. Once the available card width dropped below ~340px combined
(a laptop with the sidebar expanded, not just a phone), each column's
content (Replace-button + filename + volume slider) needed more than
its allotted ~145px, so the grid silently overflowed and was clipped by
.hg-dash-content's overflow-x:hidden - a known recurring trap in this
codebase for anything using a bare 1fr/minmax(Npx,1fr) grid without a
same-page fallback. Switched to
repeat(auto-fit,minmax(min(160px,100%),1fr)), matching the established
pattern elsewhere in this file, plus min-width:0 on the grid items and
the volume-slider row so they can actually shrink instead of forcing
overflow. Verified: single column with no overflow at 480px, 2-column
unchanged at 1400px.

+++ backend/src/modules/games/numberCalls.controller.ts
@@ -147,16 +147,22 @@ export async function uploadNumberAudio(req: Request, res: Response): Promise<vo
     const audioUrl = `/api/games/number-calls/audio-file/${filename}`;
     const numVal = parseInt(number as string, 10);
 
+    // $2::int cast is required: $2 is also used inside 'Number ' || $2 below, and Postgres
+    // infers a single type per placeholder across the whole query — left bare, that text
+    // concatenation makes it infer $2 as text, conflicting with its use as the integer
+    // `number` column ("inconsistent types deduced for parameter $2"). This 500'd on EVERY
+    // upload (both languages; EN just happened to be the one actually exercised, since NEP's
+    // existing files came from migration 042's backfill UPDATE, not this INSERT).
     const updateQuery = targetLang === 'ne'
       ? `INSERT INTO Number_Calls (number, audio_url_ne, audio_url, call_mode, call_text, default_text)
-         VALUES ($2, $1, $1, 'Audio', 'Number ' || $2, 'Number ' || $2)
+         VALUES ($2::int, $1, $1, 'Audio', 'Number ' || $2::int, 'Number ' || $2::int)
          ON CONFLICT (number) DO UPDATE
          SET audio_url_ne = EXCLUDED.audio_url_ne,
              audio_url = COALESCE(Number_Calls.audio_url_en, EXCLUDED.audio_url_ne),
              call_mode = 'Audio'
          RETURNING *`
       : `INSERT INTO Number_Calls (number, audio_url_en, audio_url, call_mode, call_text, default_text)
-         VALUES ($2, $1, $1, 'Audio', 'Number ' || $2, 'Number ' || $2)
+         VALUES ($2::int, $1, $1, 'Audio', 'Number ' || $2::int, 'Number ' || $2::int)
          ON CONFLICT (number) DO UPDATE
          SET audio_url_en = EXCLUDED.audio_url_en,
              audio_url = EXCLUDED.audio_url_en,
